% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIBS.R
\name{make.IBS.table}
\alias{make.IBS.table}
\title{Make IBS Table}
\usage{
make.IBS.table(Hapmap, ReferenceLine, OtherGenotypes, BinSize)
}
\arguments{
\item{Hapmap}{\itemize{
\item A hapmap file read into R. Needs to have chrom and pos columns named like that, but other names are flexible
}}

\item{ReferenceLine}{\itemize{
\item The line to which IBS in the others will be compared
}}

\item{OtherGenotypes}{\itemize{
\item The line or lines that will be used for comparisons
}}

\item{BinSize}{}
}
\value{

}
\description{
Make IBS Table
}
