% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIBS.R
\name{make.IBS.table}
\alias{make.IBS.table}
\title{Calculate identity by state (IBS) between two or more lines in bins across the genome}
\usage{
make.IBS.table(
  Hapmap,
  ReferenceLine,
  OtherLines,
  BinSize,
  MinBinSize = 10,
  Chromosomes
)
}
\arguments{
\item{Hapmap}{\itemize{
\item A hapmap file read into R. Needs to have chrom and pos columns named like that, but other names are flexible
}}

\item{ReferenceLine}{\itemize{
\item The line to which IBS in the others will be compared
}}

\item{OtherLines}{\itemize{
\item The line or lines that will be used for comparisons. If not specified, all other genotypes will be used
}}

\item{BinSize}{\itemize{
\item the number of SNPs per bin
}}

\item{MinBinSize}{\itemize{
\item At the ends of chromosomes or when there are many missing values, bins may be smaller than the number specified by \code{BinSize}. This number controls how small bins can be before NA is returned. (2 SNPs is probably not enough to give a good estimate of IBS, for example.)
}}

\item{Chromosomes}{\itemize{
\item A vector of which chromosomes should be used for the calculations. All are used by default
}}
}
\value{

}
\description{
Calculate identity by state (IBS) between two or more lines in bins across the genome
}
